将配置初始化从Main类中分离
